tests/wpt/tests/dom/nodes/getElementsByClassName-26.htm

<!DOCTYPE html>
<html><head>
        <title>getElementsByClassName</title>
        <meta content="multiple class attributes" name="description">
    <link href="https://dom.spec.whatwg.org/#dom-document-getelementsbyclassname" rel="help">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
    <body>
       <div id="log"></div>
         <div class="te xt">
            te xt
            <div class="te">
                te; xt
                <a class="text link" href="#foo">test link #foo</a>
            </div>
            <b class="text">text</b>
        </div>
        <div class="xt te">xt te</div>
         <script type="text/javascript">
            test(function()
            {
                var collection = document.getElementsByClassName("te xt");

                assert_equals(collection.length, 2);
                assert_equals(collection[0].parentNode.nodeName, "BODY");
                assert_equals(collection[1].parentNode.nodeName, "BODY");
            }, "multiple class attributes");
        </script>
</body></html>
